switch to the new way completely
